Move Section tests to their own file
